split msp tests
